refactoring track screen tracker readers

+++ ProtoDex/proto_dex/lib/file_manager.dart
@@ -23,14 +23,15 @@ class FileManager {
     return file;
   }
 
-  removeFile(File file) {
-    file.delete();
+  removeFile(String fileName) {
+    File file = File("${directory.path}/$fileName.json");
+    if (file.existsSync()) file.delete();
   }
 
   Future<List<File>>? findFiles(String? prefix, String? sufix) async {
     List<File> files = [];
     await for (var entity
-        in directory.list(recursive: true, followLinks: false)) {
+        in directory.list(recursive: false, followLinks: false)) {
       File file = File(entity.path);
       if (p.extension(file.path) == ".json") {
         // print("Found in:${entity.path}");
